calculator

+++ NCR_check/1.Calculator/1.Calculator/div.cpp
@@ -1,4 +1,4 @@
-float divide(int a, int b)
+int divide(int a, int b)
 {
 	extern int res;
 	res = a / b;
